+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - https://github.com/StateSmith/StateSmith/issues/81
 - nuget package - Include documentation and source code for intellisense/debugging.
   - https://github.com/StateSmith/StateSmith/issues/80
+- added `SmTransformer` pipeline that allows custom user state machine transformation!
 
 ---